Refactor login form
